+++ src/domain/stats/isLargeAccount.ts
@@ -2,12 +2,13 @@ import { IS_LARGE_ACCOUNT_KEY } from "config/localStorage";
 import { useEffect } from "react";
 import { useLocalStorageSerializeKey } from "lib/localStorage";
 import { useIsLargeAccountData } from "domain/synthetics/accountStats/useIsLargeAccountData";
-import { getStorageItem } from "lib/metrics/storage";
 
 let isLargeAccount = getIsLargeAccountStoredValue();
 
 function getIsLargeAccountStoredValue() {
-  return getStorageItem(IS_LARGE_ACCOUNT_KEY, true) === "true";
+  const storedValue = localStorage.getItem(JSON.stringify(IS_LARGE_ACCOUNT_KEY));
+
+  return storedValue === "true";
 }
 
 export function getIsLargeAccount() {
@@ -15,8 +16,8 @@ export function getIsLargeAccount() {
 }
 
 export function useIsLargeAccountTracker(account?: string) {
-  const isLargeCurrentAccount = useIsLargeAccountData(account);
-  const [isLargeAccountStoredValue, setIsLargeAccountStoredValue] = useLocalStorageSerializeKey<boolean>(
+  const isCurrentAccountLarge = useIsLargeAccountData(account);
+  const [isCurrentAccountLargeStoredValue, setIsCurrentAccountLargeStoredValue] = useLocalStorageSerializeKey<boolean>(
     IS_LARGE_ACCOUNT_KEY,
     false
   );
@@ -24,15 +25,15 @@ export function useIsLargeAccountTracker(account?: string) {
   useEffect(() => {
     if (!account) {
       isLargeAccount = false;
-    } else if (isLargeCurrentAccount !== undefined) {
-      setIsLargeAccountStoredValue(isLargeCurrentAccount);
-      isLargeAccount = isLargeCurrentAccount;
-    } else if (isLargeAccountStoredValue) {
+    } else if (isCurrentAccountLarge !== undefined) {
+      setIsCurrentAccountLargeStoredValue(isCurrentAccountLarge);
+      isLargeAccount = isCurrentAccountLarge;
+    } else if (isCurrentAccountLargeStoredValue) {
       isLargeAccount = true;
     } else {
       isLargeAccount = false;
     }
-  }, [account, isLargeCurrentAccount, isLargeAccountStoredValue, setIsLargeAccountStoredValue]);
+  }, [account, isCurrentAccountLarge, isCurrentAccountLargeStoredValue, setIsCurrentAccountLargeStoredValue]);
 
   return isLargeAccount;
 }
